refactor(broker): split proving.rs into directory layout (C35 of PR-3)
Pure mechanical move. proving.rs (1091 lines) becomes:

- proving/mod.rs (~29 lines): module decls + pub use service::ProvingService.
- proving/error.rs (~73 lines): ProvingErr enum + coded_error_impl! + the
  ProvingErr::completion_outcome helper that maps variants to
  boundless_market::telemetry::CompletionOutcome (visibility widened from
  private to pub(super) so service.rs can call it across modules).
- proving/service.rs (~1038 lines): ProvingService struct + impl,
  BrokerService impl, the set_order_failure free helper, and the original
  test module.

External path crate::proving::ProvingService preserved via the re-export.
ProvingErr stays pub but has no external consumers, so it is not re-exported
from mod.rs.
